Added unit tests and refined class diagrams
Explored the Visual Studio built-in unit testing tool and subsequently
created a C# unit testing project in the same solution. Defined a bunch
of methods to unit test initially, implemented and successfully passed
only one of the tests in order to get a feel for how difficult the
process will be.

TODO:
- Implement and document unit tests
- Document unit test results and coverage
- Combinational and state test models and test cases
Note: I don't think we need combinational tests because they are covered
by the state tests.
- Combinational and state results
- README file